pyproject.toml
[tool.poetry]
name = "learn-langchain"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
langchain = "^0.0.285"
openai = "^0.28.0"


[tool.poetry.group.dev.dependencies]
black = "^23.9.0"
pyright = "^1.1.326"
ruff = "^0.0.287"
python-dotenv = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
